% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcsqlite3-package.R
\name{adbc_driver_sqlite3}
\alias{adbc_driver_sqlite3}
\title{ADBC SQLite3 Driver}
\usage{
adbc_driver_sqlite3()
}
\value{
An \code{\link[radbc:radbc_driver_void]{radbc::radbc_driver()}}
}
\description{
ADBC SQLite3 Driver
}
\examples{
adbc_driver_sqlite3()

}
