% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcsqlite3-package.R
\docType{package}
\name{adbcsqlite3-package}
\title{adbcsqlite3: 'Arrow' Database Connectivity ('ADBC') 'SQLite3' Driver}
\description{
Provides an 'SQLite3' driver for 'Arrow' Database Connectivity ('ADBC').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/adbcsqlite3}
  \item Report bugs at \url{https://github.com/paleolimbot/adbcsqlite3/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
