% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbcsqlite3-package.R
\name{adbcsqlite3}
\alias{adbcsqlite3}
\title{ADBC SQLite3 Driver}
\usage{
adbcsqlite3()
}
\value{
An \code{\link[radbc:radbc_driver_void]{radbc::radbc_driver()}}
}
\description{
ADBC SQLite3 Driver
}
\examples{
adbcsqlite3()

}
